+++ src/umbp/src/master_client.cpp
@@ -19,7 +19,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-#include "umbp/client.h"
+#include "umbp/master_client.h"
 
 #include <grpcpp/grpcpp.h>
 #include <spdlog/spdlog.h>
